move error types to own file
