+++ src/core/mod.rs
@@ -1,5 +1,5 @@
 
-
+pub mod errortypes;
 pub mod consts;
 pub mod datatransformation;
 pub mod ethernet;
